chore(services): replace api/api.js with domain-separated services
