+++ frontend/src/services/config.js
@@ -1,0 +1,9 @@
+export const API_BASE = 'http://localhost:3250'
+
+export const getHeaders = () => {
+    const token = localStorage.getItem('token')
+    return {
+        'Content-Type': 'application/json',
+        ...(token ? { 'Authorization': `Bearer ${token}` } : {})
+    }
+}
